HUGE CHANGES - allow effective multi-category filitering
This commit adds support for multi-category filters on the website. It first populates displayed categories as a list in filter.js, and toggle addition/removal from that list when clicking on the buttons. Then filter.js checks whether the icon has matched elements as calculated in displayedCategories. As a separate store, categories are now placed in `data-categories` attributes rather than `class`.