Huge change: declarative generation of icons
This commit adds support for generating icons through the declared generate-icons.toml file, allowing customizable flavors to be geenerated. It refactors the generate-icons.py script heavily, allowing config values to be read from the TOML declaration with validated JSON schemas. The -c/--config flag is used.

To enhance reproducibility of icons across different flavors, generated SVGs are now saved in memory as variable instead of written to the newicons path. This may cause a performance issue and may be fixed later using tmp dirs.

Further additions include a --nopreserve flag to remove icons when completed, and a --checksrc flag to specify newicon paths in lieu of a config file. Docs and deps are updated.